Actividad15/src/middleware/db-connect.ts
import mongoose from "mongoose";
import { MongoMemoryServer } from "mongodb-memory-server";

let mongoServer: MongoMemoryServer;

const dbConnect = async () => {
    try {
        mongoServer = await MongoMemoryServer.create();
        const uri = mongoServer.getUri();
        await mongoose.connect(uri, { dbName: "Weather" });
        return mongoServer;
    } catch (error) {
        console.error("Error connecting to MongoDB:", error);
        throw error;
    }
};

const dbDisconnect = async () => {
    if (mongoose.connection.readyState !== 0) {
        await mongoose.disconnect();
    }
    if (mongoServer) {
        await mongoServer.stop();
    }
};

export { dbConnect, dbDisconnect };
